fix: keep version guard in b2 client
